split transactions and spans into directories
